+++ app/build.gradle
@@ -12,12 +12,14 @@ android {
         versionCode 1
         versionName "1.0"
     }
+
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+
     sourceSets {
         main.java.srcDirs += 'src/main/kotlin'
     }
@@ -28,6 +30,11 @@ kapt {
 }
 
 dependencies {
+    compile project(':domain')
+    compile project(':usecases')
+    compile project(':presentation')
+    compile project(':storage')
+
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
     compile 'com.android.support:appcompat-v7:23.1.1'
@@ -44,19 +51,21 @@ dependencies {
     kapt 'com.google.dagger:dagger-compiler:2.0'
 
     // ReQuery
-    compile 'io.requery:requery:1.0-SNAPSHOT'
+    /*compile 'io.requery:requery:1.0-SNAPSHOT'
     compile 'io.requery:requery-android:1.0-SNAPSHOT'
-    kapt 'io.requery:requery-processor:1.0-SNAPSHOT'
+    kapt 'io.requery:requery-processor:1.0-SNAPSHOT'*/
 }
+
 buildscript {
-    ext.kotlin_version = '1.0.0-rc-1036'
     repositories {
         mavenCentral()
     }
+
     dependencies {
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
+
 repositories {
     mavenCentral()
     maven { url 'http://oss.jfrog.org/artifactory/oss-snapshot-local' }
