+++ app/src/main/kotlin/io/github/dector/quotes/qoutes/QuotesModule.kt
@@ -5,26 +5,25 @@ import android.view.LayoutInflater
 import dagger.Module
 import dagger.Provides
 import io.github.dector.quotes.R
-import io.github.dector.quotes.qoutes.model.Quote
-import io.github.dector.quotes.qoutes.model.QuotesFactory
-import io.github.dector.quotes.qoutes.presentation.ColorPair
-import io.github.dector.quotes.qoutes.presentation.Palette
-import io.github.dector.quotes.qoutes.presentation.presenter.QuotesPresenter
+import io.github.dector.quotes.domain.Quote
+import io.github.dector.quotes.presentation.presenter.QuotesPresenter
 import io.github.dector.quotes.qoutes.presentation.view.QuotesView
-import io.github.dector.quotes.qoutes.storage.IStorage
 import io.github.dector.quotes.qoutes.storage.MockQuotesStorage
+import io.github.dector.quotes.storage.IStorage
+import io.github.dector.quotes.usecases.IQuotesUseCase
+import io.github.dector.quotes.usecases.QuotesUseCase
 
 @Module
 class QuotesModule() {
 
     @Provides
-    fun quotesStorage(): IStorage<Quote> = QuotesFactory(MockQuotesStorage())
+    fun quotesStorage(): IStorage<Quote?> = MockQuotesStorage()
 
     @Provides
-    fun colorsStorage(): IStorage<ColorPair> = Palette()
+    fun quotesUseCase(storage: IStorage<Quote?>): IQuotesUseCase = QuotesUseCase(storage)
 
     @Provides
-    fun quotesPresenter(quotesStorage: IStorage<Quote>, colorsStorage: IStorage<ColorPair>) = QuotesPresenter(quotesStorage, colorsStorage)
+    fun quotesPresenter(quotesUseCase: IQuotesUseCase) = QuotesPresenter(quotesUseCase)
 
     @Provides
     fun quotesView(context: Context) = QuotesView(LayoutInflater.from(context).inflate(R.layout.view_quotes, null))
