+++ app/src/main/kotlin/io/github/dector/quotes/qoutes/presentation/Palette.kt
@@ -1,16 +1,13 @@
 package io.github.dector.quotes.qoutes.presentation
 
-import io.github.dector.quotes.qoutes.presentation.Color.*
-import io.github.dector.quotes.qoutes.storage.IStorage
+import io.github.dector.quotes.qoutes.presentation.view.Color.*
+import io.github.dector.quotes.qoutes.presentation.view.ColorPair
+import io.github.dector.quotes.storage.IStorage
 import java.util.*
 
-class Palette : IStorage<ColorPair>{
+class Palette {
 
-    private val random = Random()
-
-    private var lastPair: ColorPair? = null
-
-    private val colors = arrayOf(
+    val colors = arrayOf(
             ColorPair(WHITE, RED),
             ColorPair(WHITE, PINK),
             ColorPair(WHITE, PURPLE),
@@ -31,45 +28,4 @@ class Palette : IStorage<ColorPair>{
             ColorPair(BLACK, GREY),
             ColorPair(WHITE, BLUE_GREY)
     )
-
-    override fun random(): ColorPair? {
-        var newPair = randomValue()
-        while (newPair == lastPair)
-            newPair = randomValue()
-
-        lastPair = newPair
-        return newPair
-    }
-
-    private fun randomValue() = colors[random.nextInt(colors.size)]
-}
-
-data class ColorPair(val text: Color, val background: Color)
-
-enum class Color(val value: Int) {
-    WHITE(0xffffff),
-    BLACK(0x000000),
-
-    // 500 colors
-    RED(0xf44336),
-    PINK(0xe91e63),
-    PURPLE(0x9c27b0),
-    DEEP_PURPLE(0x673ab7),
-    INDIGO(0x3f51b5),
-    BLUE(0x2196f3),
-    LIGHT_BLUE(0x03a9f4),
-    CYAN(0x00bcd4),
-    TEAL(0x009688),
-    GREEN(0x4caf50),
-    LIGHT_GREEN(0x8bc34a),
-    LIME(0xcddc39),
-    YELLOW(0xffeb3b),
-    AMBER(0xffc107),
-    ORANGE(0xff9800),
-    DEEP_ORANGE(0xff5722),
-    BROWN(0x795548),
-    GREY(0x9e9e9e),
-    BLUE_GREY(0x607d8b);
-
-    fun solidValue() = 0xff000000.toInt() or value
 }
