--- app/src/main/kotlin/io/github/dector/quotes/qoutes/presentation/presenter/IQuotesPresenter.kt
@@ -1,5 +1,0 @@
-package io.github.dector.quotes.qoutes.presentation.presenter
-
-interface IQuotesPresenter {
-
-}
