--- app/src/main/kotlin/io/github/dector/quotes/qoutes/storage/IStorage.kt
@@ -1,6 +1,0 @@
-package io.github.dector.quotes.qoutes.storage
-
-interface IStorage<Data> {
-
-    fun random(): Data?
-}
