+++ presentation/build.gradle
@@ -1,0 +1,28 @@
+apply plugin: 'kotlin'
+
+dependencies {
+    compile project(':domain')
+    compile project(':usecases')
+
+    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+}
+
+buildscript {
+    ext.kotlin_version = rootProject.ext.kotlin_version
+
+    repositories {
+        mavenCentral()
+    }
+
+    dependencies {
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+    }
+}
+
+repositories {
+    mavenCentral()
+}
+
+sourceSets {
+    main.java.srcDirs += 'src/main/kotlin'
+}
