+++ presentation/src/main/java/io/github/dector/quotes/presentation/presenter/IQuotesPresenter.kt
@@ -1,0 +1,5 @@
+package io.github.dector.quotes.presentation.presenter
+
+interface IQuotesPresenter {
+
+}
