+++ presentation/src/main/java/io/github/dector/quotes/presentation/view/IQuotesActionListener.kt
@@ -1,4 +1,4 @@
-package io.github.dector.quotes.qoutes.presentation.view
+package io.github.dector.quotes.presentation.view
 
 interface IQuotesActionListener {
