+++ presentation/src/main/java/io/github/dector/quotes/presentation/view/IQuotesView.kt
@@ -1,6 +1,6 @@
-package io.github.dector.quotes.qoutes.presentation.view
+package io.github.dector.quotes.presentation.view
 
-import io.github.dector.quotes.qoutes.presentation.Color
+import io.github.dector.quotes.qoutes.presentation.view.Color
 
 interface IQuotesView {
