+++ storage/src/main/java/io/github/dector/quotes/storage/IStorage.kt
@@ -1,0 +1,8 @@
+package io.github.dector.quotes.storage
+
+interface IStorage<Data> {
+
+    fun count(): Int
+
+    operator fun get(index: Int): Data
+}
