+++ usecases/src/main/java/io/github/dector/quotes/usecases/IQuotesUseCase.kt
@@ -1,0 +1,8 @@
+package io.github.dector.quotes.usecases
+
+import io.github.dector.quotes.domain.Quote
+
+interface IQuotesUseCase {
+
+    fun getRandomQuote(callback: (Quote?) -> Unit)
+}
